inspect archive content instead of expecting target information
One pain point when adding new archives is they need to be extracted
once and checked manually to see if they have a top dir, so that
symlinks can be correctly set.

Instead, inspect the downloaded archive and determine its top dir.
Symlinks are relative to the archive's top dir now.